mqtt-subscriber/requirements.txt
paho-mqtt
pillow
python-dotenv
numpy
supabase